Implement lazy method aliasing for mappings
This makes aliased methods nicer; first because it takes one method call
to make an alias which more clearly expresses the intent, and it makes apps trivially
faster by cutting out a method call for each alias.

There are two ways I was thinking of implementing this. The first was to keep the
custom methods contained at the cost of having to alias them for each instantiation.
The second solution was to alias the methods on the appropriate classes when they
were first used, which is faster overall but no longer contains the custom methods.

This feature is experimental, and does not currently due to MacRuby ticket #1358, but
it may have an issue or two to work out after that bug is fixed anyways.