Simplify lazy mapping loading by breaking the existing API
Instead of loading all mappings and delaying the processing until the proper
framework is loaded, HotCocoa will process the mapping right away but not load
the mapping until the proper framework is loaded.

Mappings outside of HotCocoa will need to ensure the required framework is
loaded before the mapping is defined.

Also introduce $LOADED_FRAMEWORKS, which functions like $LOADED_FEATURES but
for frameworks.