Add a benchmark for delegate building
Unfortunately, it is quite slow. I suspect the problem is that it still calls #eval with
a string, which is very expensive. It could be removed with some careful coding, but
it might be easier to rewrite the delegate builder. The only danger in a rewrite is
accidentally forgetting about a special case that used to be handled.